Refactor DNN inference: remove unused code, make features configurable
- Delete bam_schema.py (dead code, never called in production)
- Replace dnn_vcf_inference.py (1876 lines, 4 CLIs) with focused
  dnn_fold_inference.py (~300 lines, 1 CLI) keeping only the
  tensor-cache fold inference path used by the pipeline
- Remove duplicate merge/annotate code (already in dnn_merge_annotate.py)
- Remove 2 unused entry points (dnn_vcf_inference, dnn_fold_inference)
- Make DNN feature channels configurable via --channels CLI arg
  in cram_to_tensors, matching the XGBoost --features pattern
- Save channel config in tensor cache index.json metadata
- Remove tests for deleted code, update remaining imports
