+++ src/srsnv/pyproject.toml
@@ -41,9 +41,7 @@ srsnv_report = "ugbio_srsnv.srsnv_report:main"
 compare_models_report = "ugbio_srsnv.compare_models_report:main"
 cram_to_tensors = "ugbio_srsnv.deep_srsnv.cram_to_tensors:main"
 combine_splits = "ugbio_srsnv.deep_srsnv.combine_splits:main"
-dnn_vcf_inference = "ugbio_srsnv.deep_srsnv.inference.dnn_vcf_inference:main"
-dnn_fold_inference = "ugbio_srsnv.deep_srsnv.inference.dnn_vcf_inference:main_fold"
-dnn_fold_inference_from_cache = "ugbio_srsnv.deep_srsnv.inference.dnn_vcf_inference:main_fold_from_cache"
+dnn_fold_inference_from_cache = "ugbio_srsnv.deep_srsnv.inference.dnn_fold_inference:main_fold_from_cache"
 dnn_merge_and_annotate = "ugbio_srsnv.deep_srsnv.inference.dnn_merge_annotate:main"
 prepare_dnn_report = "ugbio_srsnv.prepare_dnn_report:main"
 recalibrate_dnn_folds = "ugbio_srsnv.recalibrate_dnn_folds:main"
