+++ src/srsnv/tests/unit/test_export.py
@@ -4,7 +4,6 @@ from __future__ import annotations
 
 import json
 from pathlib import Path
-from unittest.mock import MagicMock
 
 import numpy as np
 import pysam
@@ -267,96 +266,6 @@ class TestLoadInferenceEngine:
 
 
 # ---------------------------------------------------------------------------
-# Multi-GPU round robin
-# ---------------------------------------------------------------------------
-
-
-class TestMultiGPURoundRobin:
-    def test_shards_distributed_across_engines(self):
-        """Verify _predict_shard sends work to correct engine via round-robin."""
-        from ugbio_srsnv.deep_srsnv.inference.dnn_vcf_inference import _predict_shard
-
-        mock_engines = [MagicMock() for _ in range(3)]
-        for eng in mock_engines:
-            eng.predict_batch.return_value = np.array([0.5, 0.6])
-
-        chunk = {
-            "read_base_idx": torch.zeros(2, 10, dtype=torch.int16),
-            "ref_base_idx": torch.zeros(2, 10, dtype=torch.int16),
-            "tm_idx": torch.zeros(2, dtype=torch.int8),
-            "st_idx": torch.zeros(2, dtype=torch.int8),
-            "et_idx": torch.zeros(2, dtype=torch.int8),
-            "x_num_pos": torch.zeros(2, 6, 10, dtype=torch.float16),
-            "x_num_const": torch.zeros(2, 4, dtype=torch.float16),
-            "mask": torch.ones(2, 10, dtype=torch.uint8),
-            "chrom": np.array(["chr1", "chr1"], dtype=object),
-            "pos": np.array([100, 200], dtype=np.int32),
-            "rn": np.array(["read_A", "read_B"], dtype=object),
-        }
-
-        predictions = {}
-        _predict_shard(chunk, mock_engines, gpu_idx=0, batch_size=256, predictions=predictions)
-        assert mock_engines[0].predict_batch.called
-
-        predictions2 = {}
-        _predict_shard(chunk, mock_engines, gpu_idx=1, batch_size=256, predictions=predictions2)
-        assert mock_engines[1].predict_batch.called
-
-        predictions3 = {}
-        _predict_shard(chunk, mock_engines, gpu_idx=4, batch_size=256, predictions=predictions3)
-        assert mock_engines[1].predict_batch.called  # 4 % 3 == 1
-
-
-# ---------------------------------------------------------------------------
-# DNNQualAnnotator
-# ---------------------------------------------------------------------------
-
-
-class TestDNNQualAnnotator:
-    def test_annotates_ml_qual_and_filter(self, tmp_path):
-        from ugbio_srsnv.deep_srsnv.inference.dnn_vcf_inference import DNNQualAnnotator
-
-        predictions = {
-            ("chr1", 100, "read_A"): 0.99,
-            ("chr1", 200, "read_B"): 0.1,
-        }
-
-        def quality_fn(mqual):
-            return mqual * 0.9
-
-        annotator = DNNQualAnnotator(
-            predictions=predictions,
-            quality_interpolation_fn=quality_fn,
-            low_qual_threshold=10.0,
-        )
-
-        header = pysam.VariantHeader()
-        header.add_sample("SAMPLE")
-        header.add_line("##contig=<ID=chr1,length=1000>")
-        header.add_line('##FORMAT=<ID=RN,Number=1,Type=String,Description="Read name">')
-        header = annotator.edit_vcf_header(header)
-
-        assert "ML_QUAL" in header.info
-        assert "LowQual" in header.filters
-
-    def test_quality_fn_applied(self):
-        from ugbio_srsnv.deep_srsnv.inference.dnn_vcf_inference import _build_quality_fn
-
-        metadata = {"quality_recalibration_table": [[0, 10, 20, 30], [0, 8, 18, 28]]}
-        fn = _build_quality_fn(metadata)
-        assert fn is not None
-        assert fn(0) == 0
-        assert fn(10) == pytest.approx(8.0)
-        assert fn(15) == pytest.approx(13.0)
-
-    def test_no_lut_returns_none(self):
-        from ugbio_srsnv.deep_srsnv.inference.dnn_vcf_inference import _build_quality_fn
-
-        assert _build_quality_fn({}) is None
-        assert _build_quality_fn({"quality_recalibration_table": None}) is None
-
-
-# ---------------------------------------------------------------------------
 # _compose_x_num
 # ---------------------------------------------------------------------------
 
@@ -459,160 +368,3 @@ class TestProcessShardInference:
         assert stats["output_rows"] == 1
         assert stats["missing_rows"] == 0
         assert chunk["read_base_idx"].shape == (1, 64)
-
-
-# ---------------------------------------------------------------------------
-# Ensemble manifest / fold-aware inference tests
-# ---------------------------------------------------------------------------
-
-
-class TestLoadEnsembleManifest:
-    def test_valid_manifest(self, tmp_path):
-        """Valid manifest loads without error."""
-        from ugbio_srsnv.deep_srsnv.inference.dnn_vcf_inference import load_ensemble_manifest
-
-        meta0 = tmp_path / "fold0_meta.json"
-        meta1 = tmp_path / "fold1_meta.json"
-        meta0.write_text(json.dumps({"model_type": "test"}))
-        meta1.write_text(json.dumps({"model_type": "test"}))
-
-        manifest = {
-            "k_folds": 2,
-            "chrom_to_fold": {"chr1": 0, "chr2": 1},
-            "folds": [
-                {"fold_idx": 0, "metadata_path": str(meta0)},
-                {"fold_idx": 1, "metadata_path": str(meta1)},
-            ],
-        }
-        manifest_path = tmp_path / "ensemble.json"
-        manifest_path.write_text(json.dumps(manifest))
-
-        result = load_ensemble_manifest(str(manifest_path))
-        assert result["k_folds"] == 2
-        assert len(result["folds"]) == 2
-
-    def test_missing_key_raises(self, tmp_path):
-        """Manifest missing required keys raises ValueError."""
-        from ugbio_srsnv.deep_srsnv.inference.dnn_vcf_inference import load_ensemble_manifest
-
-        manifest = {"k_folds": 2, "chrom_to_fold": {"chr1": 0}}
-        manifest_path = tmp_path / "bad.json"
-        manifest_path.write_text(json.dumps(manifest))
-
-        with pytest.raises(ValueError, match="missing required key.*folds"):
-            load_ensemble_manifest(str(manifest_path))
-
-    def test_fold_count_mismatch_raises(self, tmp_path):
-        """k_folds != len(folds) raises ValueError."""
-        from ugbio_srsnv.deep_srsnv.inference.dnn_vcf_inference import load_ensemble_manifest
-
-        meta = tmp_path / "meta.json"
-        meta.write_text(json.dumps({"model_type": "test"}))
-
-        manifest = {
-            "k_folds": 3,
-            "chrom_to_fold": {"chr1": 0},
-            "folds": [{"fold_idx": 0, "metadata_path": str(meta)}],
-        }
-        manifest_path = tmp_path / "bad2.json"
-        manifest_path.write_text(json.dumps(manifest))
-
-        with pytest.raises(ValueError, match="k_folds=3 but 1"):
-            load_ensemble_manifest(str(manifest_path))
-
-    def test_missing_metadata_file_raises(self, tmp_path):
-        """Non-existent metadata_path raises FileNotFoundError."""
-        from ugbio_srsnv.deep_srsnv.inference.dnn_vcf_inference import load_ensemble_manifest
-
-        manifest = {
-            "k_folds": 1,
-            "chrom_to_fold": {"chr1": 0},
-            "folds": [{"fold_idx": 0, "metadata_path": "/nonexistent/meta.json"}],
-        }
-        manifest_path = tmp_path / "bad3.json"
-        manifest_path.write_text(json.dumps(manifest))
-
-        with pytest.raises(FileNotFoundError):
-            load_ensemble_manifest(str(manifest_path))
-
-
-class TestPartitionRowsByFold:
-    def test_basic_partitioning(self):
-        """Rows are correctly assigned to folds by chromosome."""
-        from ugbio_srsnv.deep_srsnv.inference.dnn_vcf_inference import _partition_rows_by_fold
-
-        rows = [
-            {"CHROM": "chr1", "POS": 100, "RN": "r1"},
-            {"CHROM": "chr2", "POS": 200, "RN": "r2"},
-            {"CHROM": "chr3", "POS": 300, "RN": "r3"},
-            {"CHROM": "chr1", "POS": 150, "RN": "r4"},
-        ]
-        chrom_to_fold = {"chr1": 0, "chr2": 1, "chr3": 0}
-        fold_rows, test_rows = _partition_rows_by_fold(rows, chrom_to_fold, k_folds=2)
-
-        assert len(fold_rows[0]) == 3  # chr1 x2 + chr3 x1
-        assert len(fold_rows[1]) == 1  # chr2
-        assert len(test_rows) == 0
-
-    def test_unmapped_chroms_go_to_test(self):
-        """Chromosomes not in chrom_to_fold become test reads."""
-        from ugbio_srsnv.deep_srsnv.inference.dnn_vcf_inference import _partition_rows_by_fold
-
-        rows = [
-            {"CHROM": "chr1", "POS": 100, "RN": "r1"},
-            {"CHROM": "chrX", "POS": 200, "RN": "r2"},
-            {"CHROM": "chrM", "POS": 300, "RN": "r3"},
-        ]
-        chrom_to_fold = {"chr1": 0}
-        fold_rows, test_rows = _partition_rows_by_fold(rows, chrom_to_fold, k_folds=1)
-
-        assert len(fold_rows[0]) == 1
-        assert len(test_rows) == 2
-
-    def test_empty_rows(self):
-        """Empty input produces empty outputs."""
-        from ugbio_srsnv.deep_srsnv.inference.dnn_vcf_inference import _partition_rows_by_fold
-
-        fold_rows, test_rows = _partition_rows_by_fold([], {"chr1": 0}, k_folds=1)
-        assert fold_rows == [[]]
-        assert test_rows == []
-
-
-class TestAggregateFoldProbabilities:
-    def test_single_model_passthrough(self):
-        """With K=1, aggregation returns the original probabilities."""
-        from ugbio_srsnv.deep_srsnv.inference.dnn_vcf_inference import aggregate_fold_probabilities
-
-        probs = np.array([[0.1, 0.5, 0.9]])
-        result = aggregate_fold_probabilities(probs)
-        np.testing.assert_allclose(result, [0.1, 0.5, 0.9], atol=1e-5)
-
-    def test_symmetric_average(self):
-        """Two models with symmetric probabilities around 0.5 average to ~0.5."""
-        from ugbio_srsnv.deep_srsnv.inference.dnn_vcf_inference import aggregate_fold_probabilities
-
-        probs = np.array([[0.2, 0.8], [0.8, 0.2]])
-        result = aggregate_fold_probabilities(probs)
-        np.testing.assert_allclose(result, [0.5, 0.5], atol=1e-5)
-
-    def test_logit_space_average(self):
-        """Verify aggregation is done in logit space, not probability space."""
-        from ugbio_srsnv.deep_srsnv.inference.dnn_vcf_inference import aggregate_fold_probabilities
-
-        probs = np.array([[0.1, 0.9], [0.3, 0.7]])
-        result = aggregate_fold_probabilities(probs)
-
-        eps = 1e-7
-        clipped = np.clip(probs, eps, 1 - eps)
-        logits = np.log(clipped / (1 - clipped))
-        expected = 1.0 / (1.0 + np.exp(-logits.mean(axis=0)))
-        np.testing.assert_allclose(result, expected, atol=1e-6)
-
-    def test_extreme_probabilities_clipped(self):
-        """Probabilities at 0 and 1 are clipped to avoid inf logits."""
-        from ugbio_srsnv.deep_srsnv.inference.dnn_vcf_inference import aggregate_fold_probabilities
-
-        probs = np.array([[0.0, 1.0], [0.5, 0.5]])
-        result = aggregate_fold_probabilities(probs)
-        assert np.all(np.isfinite(result))
-        assert np.all(result >= 0) and np.all(result <= 1)
