--- src/srsnv/ugbio_srsnv/deep_srsnv/inference/dnn_vcf_inference.py
@@ -1,1875 +1,0 @@
-"""High-throughput DNN inference pipeline: parallel CRAM workers, multi-GPU
-TensorRT (or PyTorch) inference, and annotated VCF output with ML_QUAL/SNVQ.
-
-Supports two modes:
-- **Single model**: ``--metadata srsnv_dnn_metadata.json``
-- **K-fold ensemble**: ``--ensemble-manifest ensemble.json`` (fold-aware routing)
-
-Usage::
-
-    dnn_vcf_inference \\
-        --featuremap-vcf input.vcf.gz \\
-        --cram source.cram \\
-        --reference ref.fa \\
-        --metadata srsnv_dnn_metadata.json \\
-        --output output.vcf.gz
-
-    dnn_vcf_inference \\
-        --featuremap-vcf input.vcf.gz \\
-        --cram source.cram \\
-        --ensemble-manifest ensemble.json \\
-        --output output.vcf.gz
-"""
-
-from __future__ import annotations
-
-import argparse
-import importlib
-import io
-import itertools
-import json
-import math
-import os
-import sys
-import tempfile
-import time
-from concurrent.futures import FIRST_COMPLETED, ProcessPoolExecutor, ThreadPoolExecutor, wait
-from dataclasses import dataclass
-from pathlib import Path
-from queue import Empty, Queue
-from threading import Event, Lock, Thread
-
-import numpy as np
-import pyarrow as pa
-import pyarrow.compute as pc
-import pysam
-from isal import igzip
-from pyarrow import parquet as pq
-from ugbio_core.logger import logger
-from ugbio_core.vcfbed.variant_annotation import VcfAnnotator
-from ugbio_featuremap.featuremap_to_dataframe import vcf_to_parquet
-
-from ugbio_srsnv.deep_srsnv.cram_to_tensors import (
-    _PARQUET_COLUMNS,
-    CHROM,
-    POS,
-    _process_shard,
-    _process_shard_from_parquet,
-    _samtools_available,
-    _worker_init,
-)
-from ugbio_srsnv.deep_srsnv.data_prep import load_vocab_config
-from ugbio_srsnv.deep_srsnv.inference.trt_engine import (
-    _compose_x_num,
-    _to_numpy,
-    _to_numpy_long,
-    load_inference_engine,
-)
-from ugbio_srsnv.srsnv_utils import MAX_PHRED, prob_to_phred
-
-ML_QUAL = "ML_QUAL"
-
-
-# ---------------------------------------------------------------------------
-# Inference configuration
-# ---------------------------------------------------------------------------
-
-
-@dataclass
-class InferenceConfig:
-    """Shared configuration for DNN inference pipeline functions."""
-
-    cram_path: str
-    reference_path: str | None = None
-    num_cram_workers: int = 4
-    shard_size: int = 10000
-    batch_size: int = 512
-    tensor_length: int = 300
-    fetch_mode: str = "samtools"
-    gpu_ids: list[int] | None = None
-    backend: str = "trt"
-    engine_path: str | None = None
-    checkpoint_path: str | None = None
-
-
-# ---------------------------------------------------------------------------
-# DNNQualAnnotator
-# ---------------------------------------------------------------------------
-
-
-class DNNQualAnnotator(VcfAnnotator):
-    """VCF annotator that stamps pre-computed DNN predictions as per-read FORMAT/MQUAL and FORMAT/SNVQ.
-
-    Matches the behaviour of the snvqual C tool used in the XGBoost pipeline:
-    each read gets its own MQUAL (Phred-scaled raw quality) and SNVQ (recalibrated),
-    and the variant-level QUAL is set to the max per-read SNVQ.
-    """
-
-    def __init__(
-        self,
-        predictions: dict,
-        quality_interpolation_fn=None,
-        low_qual_threshold: float = 40.0,
-        rn_format_key: str = "RN",
-        rn_info_key: str | None = None,
-        *,
-        _per_contig: bool = False,
-    ):
-        self.predictions = predictions
-        self.quality_fn = quality_interpolation_fn
-        self.threshold = low_qual_threshold
-        self.rn_format_key = rn_format_key
-        self.rn_info_key = rn_info_key
-        self._per_contig = _per_contig
-
-    def edit_vcf_header(self, header: pysam.VariantHeader) -> pysam.VariantHeader:
-        # FORMAT/MQUAL and FORMAT/SNVQ are already defined in the VCF from snvfind.
-        # Only add the LowQual filter definition.
-        header.filters.add("LowQual", None, None, "SNVQ below quality threshold")
-        return header
-
-    def process_records(self, records: list[pysam.VariantRecord]) -> list[pysam.VariantRecord]:
-        for rec in records:
-            rns = self._extract_rns(rec)
-            if not rns:
-                continue
-
-            # Compute per-read MQUAL and SNVQ
-            mquals: list[float] = []
-            snvqs: list[float] = []
-            for rn in rns:
-                key = (rec.pos, rn) if self._per_contig else (rec.chrom, rec.pos, rn)
-                prob = self.predictions.get(key)
-                if prob is not None and prob > 0:
-                    mq = float(prob_to_phred(np.array([prob]), max_value=MAX_PHRED)[0])
-                else:
-                    mq = 0.0
-                mquals.append(round(mq, 2))
-
-                if self.quality_fn is not None:
-                    sq = float(np.around(self.quality_fn(mq), decimals=2))
-                else:
-                    sq = mq
-                snvqs.append(round(sq, 2))
-
-            if not mquals:
-                continue
-
-            # Write per-read FORMAT fields (matching snvqual C tool)
-            rec.samples[0]["MQUAL"] = tuple(mquals)
-            rec.samples[0]["SNVQ"] = tuple(snvqs)
-
-            # Variant-level QUAL = max per-read SNVQ (matching snvqual)
-            max_snvq = max(snvqs)
-            rec.qual = max_snvq
-
-            if max_snvq >= self.threshold:
-                rec.filter.add("PASS")
-            else:
-                rec.filter.add("LowQual")
-        return records
-
-    def _extract_rns(self, rec: pysam.VariantRecord) -> list[str]:
-        """Get all read names from FORMAT/RN or INFO field.
-
-        pysam returns tuples for multi-value FORMAT fields, so we must
-        iterate the tuple elements rather than stringify the whole tuple.
-        """
-        try:
-            if rec.samples and self.rn_format_key:
-                val = rec.samples[0].get(self.rn_format_key)
-                if val is not None:
-                    if isinstance(val, tuple):
-                        return [str(v) for v in val]
-                    return [str(val)]
-        except (KeyError, IndexError):
-            pass
-        if self.rn_info_key:
-            try:
-                val = rec.info[self.rn_info_key]
-                if isinstance(val, tuple):
-                    return [str(v) for v in val]
-                return [str(val)]
-            except KeyError:
-                pass
-        return []
-
-
-# ---------------------------------------------------------------------------
-# SNVQ recalibration LUT
-# ---------------------------------------------------------------------------
-
-
-class _QualityInterpolator:
-    """Picklable MQUAL->SNVQ interpolation callable."""
-
-    def __init__(self, x_lut: np.ndarray, y_lut: np.ndarray):
-        self.x_lut = x_lut
-        self.y_lut = y_lut
-
-    def __call__(self, mqual):
-        return np.interp(mqual, self.x_lut, self.y_lut, left=0, right=self.y_lut[-1])
-
-
-def _build_quality_fn(metadata: dict) -> _QualityInterpolator | None:
-    """Build an MQUAL->SNVQ interpolation function from metadata LUT."""
-    table = metadata.get("quality_recalibration_table")
-    min_lut_entries = 2
-    if not table or len(table) < min_lut_entries:
-        return None
-    return _QualityInterpolator(np.array(table[0]), np.array(table[1]))
-
-
-# ---------------------------------------------------------------------------
-# Parallel CRAM fetch + multi-GPU inference
-# ---------------------------------------------------------------------------
-
-
-def _run_inference(
-    cram_path: str,
-    parquet_path: str,
-    metadata: dict,
-    engines: list,
-    *,
-    num_cram_workers: int = 4,
-    shard_size: int = 10000,
-    batch_size: int = 512,
-    tensor_length: int = 300,
-    reference_path: str | None = None,
-    fetch_mode: str = "samtools",
-) -> dict[tuple[str, int, str], float]:
-    """Phase 2: pipelined CRAM fetch + multi-GPU inference, returns predictions dict."""
-    encoders = load_vocab_config()
-
-    pf = pq.ParquetFile(parquet_path)
-    available_cols = [c for c in _PARQUET_COLUMNS if c in (pf.schema.names or [])]
-    table = pf.read(columns=available_cols)
-    col_dict = table.to_pydict()
-    n_rows = len(col_dict[available_cols[0]])
-    rows = [{k: col_dict[k][i] for k in col_dict} for i in range(n_rows)]
-    rows.sort(key=lambda r: (str(r.get(CHROM, "")), int(r.get(POS, 0))))
-
-    shard_common = {
-        "cram_path": cram_path,
-        "reference_path": reference_path,
-        "encoders": encoders,
-        "tensor_length": tensor_length,
-        "label": False,
-        "max_edist": None,
-        "fetch_mode": fetch_mode,
-    }
-    shard_args_list = [
-        {"shard_id": sid, "rows": rows[i : i + shard_size], **shard_common}
-        for sid, i in enumerate(range(0, len(rows), shard_size))
-    ]
-    total_shards = len(shard_args_list)
-    logger.info(
-        "Inference: %d rows -> %d shards (%d workers, %d GPUs)", n_rows, total_shards, num_cram_workers, len(engines)
-    )
-
-    t_start = time.perf_counter()
-    predictions = _pipelined_inference(shard_args_list, engines, batch_size, num_cram_workers)
-    logger.info("Inference complete: %d predictions in %.1fs", len(predictions), time.perf_counter() - t_start)
-    return predictions
-
-
-def _predict_shard(
-    chunk: dict,
-    engines: list,
-    gpu_idx: int,
-    batch_size: int,
-    predictions: dict[tuple[str, int, str], float],
-) -> None:
-    """Run a single shard's tensors through a GPU engine and collect predictions."""
-    engine = engines[gpu_idx % len(engines)]
-    n = len(chunk["chrom"])
-    if n == 0:
-        return
-    for batch_start in range(0, n, batch_size):
-        batch_end = min(batch_start + batch_size, n)
-        s = slice(batch_start, batch_end)
-        batch = {
-            "read_base_idx": chunk["read_base_idx"][s],
-            "ref_base_idx": chunk["ref_base_idx"][s],
-            "tm_idx": chunk["tm_idx"][s],
-            "st_idx": chunk["st_idx"][s],
-            "et_idx": chunk["et_idx"][s],
-            "x_num_pos": chunk["x_num_pos"][s],
-            "x_num_const": chunk["x_num_const"][s],
-            "mask": chunk["mask"][s],
-        }
-        probs = engine.predict_batch(batch)
-        for j in range(len(probs)):
-            idx = batch_start + j
-            key = (str(chunk["chrom"][idx]), int(chunk["pos"][idx]), str(chunk["rn"][idx]))
-            predictions[key] = float(probs[j])
-
-
-def _flush_preds_to_writer(
-    preds: dict[tuple[str, int, str], float],
-    writer,
-    lock: Lock,
-) -> None:
-    """Flush accumulated predictions to a ParquetWriter under a lock."""
-    table = pa.table(
-        {
-            "chrom": [k[0] for k in preds],
-            "pos": [k[1] for k in preds],
-            "rn": [k[2] for k in preds],
-            "prob": [preds[k] for k in preds],
-        }
-    )
-    with lock:
-        writer.write_table(table)
-
-
-def _gpu_consumer_thread(
-    engine,
-    queue: Queue,
-    batch_size: int,
-    local_preds: dict[tuple[str, int, str], float],
-    done_event: Event,
-    writer=None,
-    write_lock: Lock | None = None,
-    pred_count: list[int] | None = None,
-) -> None:
-    """Pull shard chunks from *queue* and run GPU inference until producers finish.
-
-    Pushes the engine's CUDA context onto this thread's stack so PyCUDA
-    operations target the correct GPU.
-
-    When *writer* is provided, predictions are flushed to parquet after each
-    shard and the local dict is cleared, keeping memory bounded.
-    """
-    engine.push_context()
-    try:
-        while True:
-            try:
-                chunk = queue.get(timeout=1.0)
-            except Empty:
-                if done_event.is_set():
-                    break
-                continue
-            if chunk is None:
-                break
-            _predict_shard(chunk, [engine], 0, batch_size, local_preds)
-            if writer is not None and local_preds:
-                _flush_preds_to_writer(local_preds, writer, write_lock)
-                if pred_count is not None:
-                    pred_count[0] += len(local_preds)
-                local_preds.clear()
-            queue.task_done()
-    finally:
-        engine.pop_context()
-
-
-def _produce_shards_serial(
-    shard_args_list: list[dict],
-    fn,
-    q: Queue,
-    total_shards: int,
-    t_start: float,
-) -> None:
-    """Feed shard chunks into *q* sequentially (single-worker mode)."""
-    for completed, shard_args in enumerate(shard_args_list, 1):
-        _sid, chunk, _stats = fn(**shard_args)
-        q.put(chunk)
-        if completed == 1 or completed % 10 == 0 or completed == total_shards:
-            logger.info(
-                "Inference progress: %d/%d shards (%.1fs)",
-                completed,
-                total_shards,
-                time.perf_counter() - t_start,
-            )
-
-
-def _produce_shards_parallel(
-    shard_args_list: list[dict],
-    fn,
-    q: Queue,
-    total_shards: int,
-    effective_workers: int,
-    t_start: float,
-) -> None:
-    """Feed shard chunks into *q* using a process pool."""
-    completed = 0
-    with ProcessPoolExecutor(max_workers=effective_workers, initializer=_worker_init) as pool:
-        max_pending = effective_workers * 2
-        shard_iter = iter(shard_args_list)
-        pending = {pool.submit(fn, **sa) for sa in itertools.islice(shard_iter, max_pending)}
-        while pending:
-            done, pending = wait(pending, return_when=FIRST_COMPLETED)
-            for future in done:
-                _sid, chunk, _stats = future.result()
-                q.put(chunk)
-                completed += 1
-                if completed == 1 or completed % 10 == 0 or completed == total_shards:
-                    logger.info(
-                        "Inference progress: %d/%d shards (%.1fs)",
-                        completed,
-                        total_shards,
-                        time.perf_counter() - t_start,
-                    )
-            for sa in itertools.islice(shard_iter, len(done)):
-                pending.add(pool.submit(fn, **sa))
-
-
-def _start_gpu_consumers(
-    engines: list,
-    q: Queue,
-    batch_size: int,
-    done_event: Event,
-    local_preds: list[dict],
-    *,
-    writer=None,
-    write_lock: Lock | None = None,
-    pred_counts: list[list[int]] | None = None,
-) -> list[Thread]:
-    """Start one GPU consumer thread per engine and return thread handles."""
-    gpu_threads: list[Thread] = []
-    streaming = writer is not None
-    for i, engine in enumerate(engines):
-        t = Thread(
-            target=_gpu_consumer_thread,
-            args=(engine, q, batch_size, local_preds[i], done_event),
-            kwargs={
-                "writer": writer,
-                "write_lock": write_lock,
-                "pred_count": pred_counts[i] if streaming else None,
-            },
-            daemon=True,
-        )
-        t.start()
-        gpu_threads.append(t)
-    return gpu_threads
-
-
-def _pipelined_inference(
-    shard_args_list: list[dict],
-    engines: list,
-    batch_size: int,
-    num_cram_workers: int,
-    process_fn=None,
-    output_path: str | None = None,
-) -> dict[tuple[str, int, str], float] | int:
-    """Run CPU shard processing and multi-GPU inference concurrently.
-
-    CPU workers produce tensor chunks via ``ProcessPoolExecutor`` and push them
-    into a ``Queue``.  Per-GPU consumer threads pull from the queue and run
-    ``predict_batch`` in parallel, giving true multi-GPU utilisation.
-
-    Parameters
-    ----------
-    shard_args_list
-        List of keyword-argument dicts, one per shard.  Each dict is unpacked
-        as ``process_fn(**shard_args)``.
-    process_fn
-        Callable that processes a shard.  Defaults to ``_process_shard``.
-    output_path
-        If provided, predictions are written incrementally to this parquet file
-        (one row group per shard) and the total count is returned instead of a
-        dict.  This keeps memory bounded to one shard per GPU thread.
-    """
-
-    fn = process_fn or _process_shard
-    total_shards = len(shard_args_list)
-    effective_workers = max(1, num_cram_workers)
-
-    prefetch = len(engines) * 2
-    q: Queue = Queue(maxsize=prefetch)
-    done_event = Event()
-    local_preds: list[dict[tuple[str, int, str], float]] = [{} for _ in engines]
-
-    streaming = output_path is not None
-    writer = None
-    write_lock = Lock() if streaming else None
-    pred_counts: list[list[int]] = [[0] for _ in engines] if streaming else []
-
-    if streaming:
-        schema = pa.schema(
-            [
-                ("chrom", pa.string()),
-                ("pos", pa.int64()),
-                ("rn", pa.string()),
-                ("prob", pa.float64()),
-            ]
-        )
-        writer = pq.ParquetWriter(output_path, schema)
-
-    gpu_threads = _start_gpu_consumers(
-        engines,
-        q,
-        batch_size,
-        done_event,
-        local_preds,
-        writer=writer,
-        write_lock=write_lock,
-        pred_counts=pred_counts if streaming else None,
-    )
-
-    t_start = time.perf_counter()
-
-    try:
-        if effective_workers == 1:
-            _produce_shards_serial(shard_args_list, fn, q, total_shards, t_start)
-        else:
-            _produce_shards_parallel(shard_args_list, fn, q, total_shards, effective_workers, t_start)
-    finally:
-        done_event.set()
-        for t in gpu_threads:
-            t.join()
-        if writer is not None:
-            writer.close()
-
-    if streaming:
-        return sum(c[0] for c in pred_counts)
-
-    predictions: dict[tuple[str, int, str], float] = {}
-    for lp in local_preds:
-        predictions.update(lp)
-
-    return predictions
-
-
-# ---------------------------------------------------------------------------
-# Ensemble manifest (k-fold)
-# ---------------------------------------------------------------------------
-
-
-def load_ensemble_manifest(path: str) -> dict:
-    """Load and validate an ensemble manifest JSON for k-fold inference.
-
-    Expected schema::
-
-        {
-            "k_folds": 3,
-            "chrom_to_fold": {"chr1": 0, "chr2": 1, ...},
-            "folds": [
-                {"fold_idx": 0, "metadata_path": "/path/to/fold0/metadata.json"},
-                ...
-            ],
-            "quality_recalibration_table": [[...], [...]]  // optional
-        }
-    """
-    with open(path) as f:
-        manifest = json.load(f)
-
-    for key in ("k_folds", "chrom_to_fold", "folds"):
-        if key not in manifest:
-            raise ValueError(f"Ensemble manifest missing required key: {key!r}")
-
-    k_folds = manifest["k_folds"]
-    folds = manifest["folds"]
-    if len(folds) != k_folds:
-        raise ValueError(f"k_folds={k_folds} but {len(folds)} fold entries provided")
-
-    for i, fold in enumerate(folds):
-        if "fold_idx" not in fold or "metadata_path" not in fold:
-            raise ValueError(f"Fold entry {i} missing 'fold_idx' or 'metadata_path'")
-        meta_p = Path(fold["metadata_path"])
-        if not meta_p.exists():
-            raise FileNotFoundError(f"Fold {fold['fold_idx']} metadata not found: {meta_p}")
-
-    return manifest
-
-
-def _partition_rows_by_fold(
-    rows: list[dict],
-    chrom_to_fold: dict[str, int],
-    k_folds: int,
-) -> tuple[list[list[dict]], list[dict]]:
-    """Partition rows into per-fold groups and a test group.
-
-    Returns ``(fold_rows, test_rows)`` where ``fold_rows[k]`` contains rows
-    for fold k and ``test_rows`` contains rows whose chromosome is not in
-    ``chrom_to_fold`` (predicted by all models and aggregated).
-    """
-    fold_rows: list[list[dict]] = [[] for _ in range(k_folds)]
-    test_rows: list[dict] = []
-    for row in rows:
-        chrom = str(row.get(CHROM, row.get("chrom", "")))
-        fold_id = chrom_to_fold.get(chrom)
-        if fold_id is not None and 0 <= fold_id < k_folds:
-            fold_rows[fold_id].append(row)
-        else:
-            test_rows.append(row)
-    return fold_rows, test_rows
-
-
-def aggregate_fold_probabilities(prob_matrix: np.ndarray) -> np.ndarray:
-    """Average K models' probabilities in logit space.
-
-    Parameters
-    ----------
-    prob_matrix
-        Shape ``(K, N)`` array of probabilities from K fold models for N reads.
-
-    Returns
-    -------
-    np.ndarray
-        Shape ``(N,)`` aggregated probabilities.
-    """
-    eps = 1e-7
-    clipped = np.clip(prob_matrix, eps, 1.0 - eps)
-    logits = np.log(clipped / (1.0 - clipped))
-    mean_logit = np.mean(logits, axis=0)
-    return 1.0 / (1.0 + np.exp(-mean_logit))
-
-
-# ---------------------------------------------------------------------------
-# Fold-aware inference (k-fold ensemble)
-# ---------------------------------------------------------------------------
-
-
-def _run_fold_inference(
-    config: InferenceConfig,
-    parquet_path: str,
-    n_rows: int,
-    columns: list[str],
-    engines: list,
-    *,
-    output_path: str | None = None,
-) -> dict[tuple[str, int, str], float] | int:
-    """Run inference from a pre-sorted parquet file with row-group-per-shard layout.
-
-    Each worker reads its own row group from disk — no row dicts in the main process.
-
-    When *output_path* is provided, predictions are streamed to parquet
-    incrementally and the total count is returned instead of a dict.
-    """
-    if n_rows == 0:
-        return 0 if output_path is not None else {}
-
-    encoders = load_vocab_config()
-
-    pf = pq.ParquetFile(parquet_path)
-    total_shards = pf.metadata.num_row_groups
-
-    shard_common = {
-        "cram_path": config.cram_path,
-        "reference_path": config.reference_path,
-        "encoders": encoders,
-        "tensor_length": config.tensor_length,
-        "label": False,
-        "max_edist": None,
-        "fetch_mode": config.fetch_mode,
-    }
-
-    shard_args_list = [
-        {"shard_id": rg, "parquet_path": parquet_path, "row_group_id": rg, "columns": columns, **shard_common}
-        for rg in range(total_shards)
-    ]
-
-    logger.info(
-        "Fold inference: %d rows -> %d row-group shards (%d workers, %d GPUs)",
-        n_rows,
-        total_shards,
-        config.num_cram_workers,
-        len(engines),
-    )
-
-    t_start = time.perf_counter()
-    result = _pipelined_inference(
-        shard_args_list,
-        engines,
-        config.batch_size,
-        config.num_cram_workers,
-        process_fn=_process_shard_from_parquet,
-        output_path=output_path,
-    )
-    elapsed = time.perf_counter() - t_start
-
-    n_preds = result if isinstance(result, int) else len(result)
-    logger.info(
-        "Fold inference: %d rows, %d shards, %d predictions in %.1fs",
-        n_rows,
-        total_shards,
-        n_preds,
-        elapsed,
-    )
-    return result
-
-
-def _run_fold_inference_from_rows(
-    cram_path: str,
-    rows: list[dict],
-    engines: list,
-    *,
-    num_cram_workers: int = 4,
-    shard_size: int = 10000,
-    batch_size: int = 512,
-    tensor_length: int = 300,
-    reference_path: str | None = None,
-    fetch_mode: str = "samtools",
-) -> dict[tuple[str, int, str], float]:
-    """Run inference on pre-loaded rows (used by ensemble path).
-
-    Like ``_run_fold_inference`` but accepts in-memory rows instead of a parquet path.
-    """
-    if not rows:
-        return {}
-
-    encoders = load_vocab_config()
-
-    rows_sorted = sorted(rows, key=lambda r: (str(r.get(CHROM, "")), int(r.get(POS, 0))))
-    shard_common = {
-        "cram_path": cram_path,
-        "reference_path": reference_path,
-        "encoders": encoders,
-        "tensor_length": tensor_length,
-        "label": False,
-        "max_edist": None,
-        "fetch_mode": fetch_mode,
-    }
-    shard_args_list = [
-        {"shard_id": sid, "rows": rows_sorted[i : i + shard_size], **shard_common}
-        for sid, i in enumerate(range(0, len(rows_sorted), shard_size))
-    ]
-    total_shards = len(shard_args_list)
-
-    t_start = time.perf_counter()
-    predictions = _pipelined_inference(shard_args_list, engines, batch_size, num_cram_workers)
-    elapsed = time.perf_counter() - t_start
-
-    logger.info(
-        "Fold inference: %d rows, %d shards, %d predictions in %.1fs",
-        len(rows),
-        total_shards,
-        len(predictions),
-        elapsed,
-    )
-    return predictions
-
-
-def _run_ensemble_inference(
-    manifest: dict,
-    config: InferenceConfig,
-    parquet_path: str,
-) -> dict[tuple[str, int, str], float]:
-    """Fold-aware inference: iterate folds sequentially, all GPUs data-parallel per fold."""
-
-    k_folds = manifest["k_folds"]
-    chrom_to_fold = manifest["chrom_to_fold"]
-    fold_entries = sorted(manifest["folds"], key=lambda f: f["fold_idx"])
-
-    pf = pq.ParquetFile(parquet_path)
-    available_cols = [c for c in _PARQUET_COLUMNS if c in (pf.schema.names or [])]
-    table = pf.read(columns=available_cols)
-    col_dict = table.to_pydict()
-    n_rows = len(col_dict[available_cols[0]])
-    all_rows = [{k: col_dict[k][i] for k in col_dict} for i in range(n_rows)]
-
-    fold_rows, test_rows = _partition_rows_by_fold(all_rows, chrom_to_fold, k_folds)
-    fold_sizes = [len(fr) for fr in fold_rows]
-    logger.info(
-        "Ensemble: %d total rows -> folds %s + %d test rows",
-        n_rows,
-        fold_sizes,
-        len(test_rows),
-    )
-
-    all_predictions: dict[tuple[str, int, str], float] = {}
-    t_start = time.perf_counter()
-
-    infer_kwargs = {
-        "cram_path": config.cram_path,
-        "num_cram_workers": config.num_cram_workers,
-        "shard_size": config.shard_size,
-        "batch_size": config.batch_size,
-        "tensor_length": config.tensor_length,
-        "reference_path": config.reference_path,
-        "fetch_mode": config.fetch_mode,
-    }
-
-    for fold_entry in fold_entries:
-        fold_k = fold_entry["fold_idx"]
-        meta_path = fold_entry["metadata_path"]
-        rows_k = fold_rows[fold_k] if fold_k < len(fold_rows) else []
-        if not rows_k:
-            logger.info("Fold %d: 0 reads, skipping", fold_k)
-            continue
-
-        logger.info("Fold %d: loading model from %s (%d reads)", fold_k, meta_path, len(rows_k))
-        engines = _create_engines(meta_path, backend=config.backend, gpu_ids=config.gpu_ids)
-        try:
-            fold_preds = _run_fold_inference_from_rows(rows=rows_k, engines=engines, **infer_kwargs)
-            all_predictions.update(fold_preds)
-        finally:
-            for eng in engines:
-                eng.close()
-
-    if test_rows:
-        logger.info("Processing %d test reads (aggregating across %d fold models)", len(test_rows), k_folds)
-        test_keys = [
-            (str(r.get(CHROM, r.get("chrom", ""))), int(r.get(POS, r.get("pos", 0))), str(r.get("RN", r.get("rn", ""))))
-            for r in test_rows
-        ]
-        prob_matrix = np.zeros((k_folds, len(test_rows)), dtype=np.float64)
-
-        for fold_entry in fold_entries:
-            fold_k = fold_entry["fold_idx"]
-            meta_path = fold_entry["metadata_path"]
-            logger.info("Test reads: predicting with fold %d model", fold_k)
-            engines = _create_engines(meta_path, backend=config.backend, gpu_ids=config.gpu_ids)
-            try:
-                test_preds = _run_fold_inference_from_rows(rows=test_rows, engines=engines, **infer_kwargs)
-                for j, key in enumerate(test_keys):
-                    prob_matrix[fold_k, j] = test_preds.get(key, 0.5)
-            finally:
-                for eng in engines:
-                    eng.close()
-
-        aggregated = aggregate_fold_probabilities(prob_matrix)
-        for j, key in enumerate(test_keys):
-            all_predictions[key] = float(aggregated[j])
-        logger.info("Test reads aggregated: %d predictions", len(test_keys))
-
-    elapsed = time.perf_counter() - t_start
-    logger.info("Ensemble inference complete: %d predictions in %.1fs", len(all_predictions), elapsed)
-    return all_predictions
-
-
-def _create_engines(
-    metadata_path: str,
-    *,
-    backend: str = "pytorch",
-    gpu_ids: list[int],
-) -> list:
-    """Load one inference engine per GPU for a given metadata file.
-
-    For TRT engines, each engine's CUDA context is popped after creation so it
-    can be pushed onto consumer threads later.
-    """
-    engines = []
-    for gid in gpu_ids:
-        try:
-            eng = load_inference_engine(metadata_path, backend=backend, device_id=gid)
-            if hasattr(eng, "pop_context"):
-                eng.pop_context()
-            engines.append(eng)
-        except Exception:
-            logger.warning("Failed to create engine on GPU:%d for %s, skipping", gid, metadata_path, exc_info=True)
-    if not engines:
-        raise RuntimeError(f"No inference engines could be created from {metadata_path}")
-    return engines
-
-
-# ---------------------------------------------------------------------------
-# Main pipeline
-# ---------------------------------------------------------------------------
-
-
-def _resolve_gpu_ids(gpu_ids: list[int] | None) -> list[int]:
-    """Return explicit GPU IDs, detecting available devices if needed."""
-    if gpu_ids is not None:
-        return gpu_ids
-    _torch = importlib.import_module("torch")
-
-    if _torch.cuda.is_available():
-        return list(range(_torch.cuda.device_count()))
-    return [0]
-
-
-def _resolve_fetch_mode(fetch_mode: str) -> str:
-    """Validate fetch_mode, falling back to pysam if samtools is unavailable."""
-    if fetch_mode == "samtools" and not _samtools_available():
-        logger.warning("samtools not found; falling back to pysam fetch mode")
-        return "pysam"
-    return fetch_mode
-
-
-def _ensure_parquet(featuremap_vcf: str, parquet_path: str | None) -> str:
-    """Return an existing parquet path or convert the VCF to a new one."""
-    if parquet_path is not None:
-        return parquet_path
-    tmp_dir = tempfile.mkdtemp(prefix="dnn_infer_")
-    out_path = str(Path(tmp_dir) / "featuremap.parquet")
-    logger.info("Converting featuremap VCF to parquet: %s", out_path)
-    _vcf_to_parquet(featuremap_vcf, out_path)
-    return out_path
-
-
-def _run_single_model_inference(
-    metadata_path: str,
-    config: InferenceConfig,
-    parquet_path: str,
-) -> tuple[dict[tuple[str, int, str], float], _QualityInterpolator | None]:
-    """Load a single model, run inference, return predictions and quality fn."""
-    with open(metadata_path) as f:
-        metadata = json.load(f)
-    quality_fn = _build_quality_fn(metadata)
-
-    engines = []
-    for gid in config.gpu_ids:
-        try:
-            eng = load_inference_engine(
-                metadata_path,
-                backend=config.backend,
-                device_id=gid,
-                engine_path=config.engine_path,
-                checkpoint_path=config.checkpoint_path,
-            )
-            if hasattr(eng, "pop_context"):
-                eng.pop_context()
-            engines.append(eng)
-        except Exception:
-            logger.warning("Failed to create engine on GPU:%d, skipping", gid, exc_info=True)
-    if not engines:
-        raise RuntimeError("No inference engines could be created")
-    logger.info("Created %d %s engine(s) on GPUs %s", len(engines), config.backend, config.gpu_ids[: len(engines)])
-
-    try:
-        predictions = _run_inference(
-            cram_path=config.cram_path,
-            parquet_path=parquet_path,
-            metadata=metadata,
-            engines=engines,
-            num_cram_workers=config.num_cram_workers,
-            shard_size=config.shard_size,
-            batch_size=config.batch_size,
-            tensor_length=config.tensor_length,
-            reference_path=config.reference_path,
-            fetch_mode=config.fetch_mode,
-        )
-    finally:
-        for eng in engines:
-            eng.close()
-
-    return predictions, quality_fn
-
-
-def run_inference_pipeline(
-    featuremap_vcf: str,
-    config: InferenceConfig,
-    output_vcf: str,
-    *,
-    metadata_path: str | None = None,
-    ensemble_manifest_path: str | None = None,
-    low_qual_threshold: float = 40.0,
-    parquet_path: str | None = None,
-) -> str:
-    """Run the full DNN VCF inference pipeline.
-
-    Supports two modes:
-    - **Single model**: provide ``metadata_path``
-    - **K-fold ensemble**: provide ``ensemble_manifest_path``
-
-    Returns the path to the output VCF.
-    """
-    if not metadata_path and not ensemble_manifest_path:
-        raise ValueError("Either metadata_path or ensemble_manifest_path must be provided")
-
-    torch_mp = importlib.import_module("torch.multiprocessing")
-
-    torch_mp.set_sharing_strategy("file_system")
-
-    t0 = time.perf_counter()
-
-    config = InferenceConfig(
-        cram_path=config.cram_path,
-        reference_path=config.reference_path,
-        num_cram_workers=config.num_cram_workers,
-        shard_size=config.shard_size,
-        batch_size=config.batch_size,
-        tensor_length=config.tensor_length,
-        fetch_mode=_resolve_fetch_mode(config.fetch_mode),
-        gpu_ids=_resolve_gpu_ids(config.gpu_ids),
-        backend=config.backend,
-        engine_path=config.engine_path,
-        checkpoint_path=config.checkpoint_path,
-    )
-    parquet_path = _ensure_parquet(featuremap_vcf, parquet_path)
-
-    # ── Dispatch: single model vs k-fold ensemble ──
-    if ensemble_manifest_path:
-        manifest = load_ensemble_manifest(ensemble_manifest_path)
-        quality_fn = _build_quality_fn(manifest)
-        predictions = _run_ensemble_inference(
-            manifest=manifest,
-            config=config,
-            parquet_path=parquet_path,
-        )
-    else:
-        predictions, quality_fn = _run_single_model_inference(
-            metadata_path,
-            config,
-            parquet_path,
-        )
-
-    # ── Phase 3: VCF annotation ──
-    annotator = DNNQualAnnotator(
-        predictions=predictions,
-        quality_interpolation_fn=quality_fn,
-        low_qual_threshold=low_qual_threshold,
-    )
-    VcfAnnotator.process_vcf(
-        annotators=[annotator],
-        input_path=featuremap_vcf,
-        output_path=output_vcf,
-    )
-
-    wall = time.perf_counter() - t0
-    logger.info("DNN VCF inference complete in %.1fs: %d predictions -> %s", wall, len(predictions), output_vcf)
-    return output_vcf
-
-
-def _vcf_to_parquet(vcf_path: str, parquet_path: str) -> None:
-    """Lightweight VCF->parquet for inference (extracts CHROM, POS, RN, etc.)."""
-    vcf_to_parquet(vcf_path, parquet_path, drop_format={"AD"})
-
-
-# ---------------------------------------------------------------------------
-# CLI
-# ---------------------------------------------------------------------------
-
-
-def _parse_args(argv: list[str] | None = None) -> argparse.Namespace:
-    ap = argparse.ArgumentParser(
-        description="DNN VCF inference: CRAM -> TensorRT/PyTorch -> annotated VCF",
-    )
-    ap.add_argument("--featuremap-vcf", required=True, help="Input featuremap VCF")
-    ap.add_argument("--cram", required=True, help="Source CRAM file")
-    ap.add_argument("--reference", default=None, help="Reference FASTA for CRAM decoding")
-    ap.add_argument("--output", required=True, help="Output annotated VCF path")
-    ap.add_argument("--parquet", default=None, help="Pre-computed featuremap parquet (skip VCF conversion)")
-
-    model_group = ap.add_mutually_exclusive_group(required=True)
-    model_group.add_argument("--metadata", default=None, help="Path to srsnv_dnn_metadata.json (single model)")
-    model_group.add_argument("--ensemble-manifest", default=None, help="Path to ensemble manifest JSON (k-fold)")
-
-    ap.add_argument("--engine-path", default=None, help="Override TRT engine path (single model only)")
-    ap.add_argument("--checkpoint", default=None, help="Override model checkpoint path (single model only)")
-    ap.add_argument("--backend", default="trt", choices=["trt", "pytorch"], help="Inference backend (default: trt)")
-    ap.add_argument("--gpus", default=None, help="Comma-separated GPU IDs (default: all visible)")
-    ap.add_argument("--num-cram-workers", type=int, default=None, help="CPU workers for CRAM fetch (default: all CPUs)")
-    ap.add_argument("--shard-size", type=int, default=10000, help="Rows per shard")
-    ap.add_argument("--batch-size", type=int, default=512, help="GPU inference batch size")
-    ap.add_argument("--tensor-length", type=int, default=300, help="Padded sequence length")
-    ap.add_argument("--low-qual-threshold", type=float, default=40.0, help="SNVQ threshold for PASS filter")
-    ap.add_argument("--fetch-mode", default="samtools", choices=["samtools", "pysam"], help="CRAM fetch strategy")
-    return ap.parse_args(argv)
-
-
-def run(argv: list[str] | None = None) -> None:
-    if argv is None:
-        argv = sys.argv[1:]
-    args = _parse_args(argv)
-
-    gpu_ids = None
-    if args.gpus:
-        gpu_ids = [int(g.strip()) for g in args.gpus.split(",")]
-
-    num_workers = args.num_cram_workers
-    if num_workers is None:
-        num_workers = os.cpu_count() or 1
-
-    config = InferenceConfig(
-        cram_path=args.cram,
-        reference_path=args.reference,
-        num_cram_workers=num_workers,
-        shard_size=args.shard_size,
-        batch_size=args.batch_size,
-        tensor_length=args.tensor_length,
-        fetch_mode=args.fetch_mode,
-        gpu_ids=gpu_ids,
-        backend=args.backend,
-        engine_path=args.engine_path,
-        checkpoint_path=args.checkpoint,
-    )
-
-    run_inference_pipeline(
-        featuremap_vcf=args.featuremap_vcf,
-        config=config,
-        output_vcf=args.output,
-        metadata_path=args.metadata,
-        ensemble_manifest_path=args.ensemble_manifest,
-        low_qual_threshold=args.low_qual_threshold,
-        parquet_path=args.parquet,
-    )
-
-
-def main() -> None:
-    run()
-
-
-# ---------------------------------------------------------------------------
-# Per-fold inference CLI  (dnn_fold_inference)
-# ---------------------------------------------------------------------------
-
-
-def _parse_fold_args(argv: list[str] | None = None) -> argparse.Namespace:
-    ap = argparse.ArgumentParser(
-        description="DNN single-fold inference: predict one fold's chromosomes",
-    )
-    ap.add_argument("--featuremap-vcf", required=True, help="Input featuremap VCF")
-    ap.add_argument("--cram", required=True, help="Source CRAM file")
-    ap.add_argument("--reference", default=None, help="Reference FASTA for CRAM decoding")
-    ap.add_argument("--fold-metadata", required=True, help="Path to this fold's srsnv_dnn_metadata.json")
-    ap.add_argument("--split-manifest", required=True, help="Path to split_manifest.json (chrom->fold mapping)")
-    ap.add_argument("--fold-idx", type=int, required=True, help="Fold index (0-based)")
-    ap.add_argument("--output", required=True, help="Output predictions parquet path")
-    ap.add_argument("--parquet", default=None, help="Pre-computed featuremap parquet (skip VCF conversion)")
-    ap.add_argument("--backend", default="trt", choices=["trt", "pytorch"], help="Inference backend")
-    ap.add_argument("--gpus", default=None, help="Comma-separated GPU IDs (default: all visible)")
-    ap.add_argument("--num-cram-workers", type=int, default=None, help="CPU workers for CRAM fetch")
-    ap.add_argument("--shard-size", type=int, default=10000)
-    ap.add_argument("--batch-size", type=int, default=512)
-    ap.add_argument("--tensor-length", type=int, default=300)
-    ap.add_argument("--fetch-mode", default="samtools", choices=["samtools", "pysam"])
-    return ap.parse_args(argv)
-
-
-def _resolve_fold_target_chroms(
-    parquet_file: pq.ParquetFile,
-    chrom_to_fold: dict[str, int],
-    fold_idx: int,
-) -> set[str]:
-    """Determine which chromosomes this fold should process.
-
-    Includes fold-assigned chroms plus (for fold 0) any chroms not in
-    the mapping (test chromosomes).
-    """
-    my_chroms: set[str] = {chrom for chrom, fid in chrom_to_fold.items() if fid == fold_idx}
-    test_chroms: set[str] = set()
-    if fold_idx == 0:
-        chrom_col_arr = parquet_file.read(columns=[CHROM]).column(CHROM)
-        all_chroms_in_data = set(chrom_col_arr.to_pylist())
-        test_chroms = all_chroms_in_data - set(chrom_to_fold.keys())
-        del chrom_col_arr
-    return my_chroms | test_chroms
-
-
-def _filter_parquet_by_chroms(
-    parquet_file: pq.ParquetFile,
-    available_cols: list[str],
-    target_chroms: set[str],
-) -> pa.Table:
-    """Read parquet row-groups and return only rows matching *target_chroms*."""
-    batches = []
-    for batch in parquet_file.iter_batches(batch_size=500_000, columns=available_cols):
-        chrom_arr = batch.column(CHROM)
-        mask = pa.compute.is_in(chrom_arr, value_set=pa.array(list(target_chroms)))
-        filtered_batch = batch.filter(mask)
-        if len(filtered_batch) > 0:
-            batches.append(filtered_batch)
-    if batches:
-        return pa.concat_tables([pa.Table.from_batches([b]) for b in batches])
-    return pa.table({c: [] for c in available_cols})
-
-
-def _parse_gpu_ids_from_arg(gpus_arg: str | None) -> list[int]:
-    """Parse a comma-separated GPU IDs string or auto-detect available GPUs."""
-    if gpus_arg:
-        return [int(g.strip()) for g in gpus_arg.split(",")]
-    return _resolve_gpu_ids(None)
-
-
-def run_fold(argv: list[str] | None = None) -> None:
-    """Run inference for a single fold's chromosomes and write predictions parquet."""
-    torch_mp = importlib.import_module("torch.multiprocessing")
-
-    torch_mp.set_sharing_strategy("file_system")
-
-    if argv is None:
-        argv = sys.argv[1:]
-    args = _parse_fold_args(argv)
-
-    fold_idx = args.fold_idx
-    t0 = time.perf_counter()
-
-    gpu_ids = _parse_gpu_ids_from_arg(args.gpus)
-    num_workers = args.num_cram_workers if args.num_cram_workers is not None else max(1, (os.cpu_count() or 4) - 2)
-    fetch_mode = _resolve_fetch_mode(args.fetch_mode)
-    parquet_path = _ensure_parquet(args.featuremap_vcf, args.parquet)
-
-    # Read split manifest
-    with open(args.split_manifest) as f:
-        manifest = json.load(f)
-    chrom_to_fold: dict[str, int] = manifest.get("chrom_to_fold", {})
-
-    # Filter parquet to this fold's chromosomes
-    pf = pq.ParquetFile(parquet_path)
-    available_cols = [c for c in _PARQUET_COLUMNS if c in (pf.schema.names or [])]
-    target_chroms = _resolve_fold_target_chroms(pf, chrom_to_fold, fold_idx)
-    filtered = _filter_parquet_by_chroms(pf, available_cols, target_chroms)
-
-    n_rows = len(filtered)
-    logger.info("Fold %d: %d rows (total %d, filtered %d)", fold_idx, n_rows, pf.metadata.num_rows, n_rows)
-
-    if n_rows == 0:
-        logger.info("Fold %d: no rows, writing empty predictions", fold_idx)
-        pq.write_table(pa.table({"chrom": [], "pos": [], "rn": [], "prob": []}), args.output)
-        return
-
-    # Sort and write to sharded parquet
-    sort_indices = pa.compute.sort_indices(filtered, sort_keys=[(CHROM, "ascending"), (POS, "ascending")])
-    sorted_table = filtered.take(sort_indices)
-    del filtered, sort_indices
-
-    fold_parquet = str(Path(args.output).parent / f"fold_{fold_idx}_sorted.parquet")
-    available_cols_final = [c for c in available_cols if c in sorted_table.column_names]
-    pq.write_table(sorted_table, fold_parquet, row_group_size=args.shard_size)
-    del sorted_table
-
-    # Run inference
-    engines = _create_engines(args.fold_metadata, backend=args.backend, gpu_ids=gpu_ids)
-    logger.info("Fold %d: %d %s engine(s) on GPUs %s", fold_idx, len(engines), args.backend, gpu_ids[: len(engines)])
-
-    fold_config = InferenceConfig(
-        cram_path=args.cram,
-        reference_path=args.reference,
-        num_cram_workers=num_workers,
-        shard_size=args.shard_size,
-        batch_size=args.batch_size,
-        tensor_length=args.tensor_length,
-        fetch_mode=fetch_mode,
-        gpu_ids=gpu_ids,
-        backend=args.backend,
-    )
-
-    try:
-        n_preds = _run_fold_inference(
-            config=fold_config,
-            parquet_path=fold_parquet,
-            n_rows=n_rows,
-            columns=available_cols_final,
-            engines=engines,
-            output_path=args.output,
-        )
-    finally:
-        for eng in engines:
-            eng.close()
-
-    elapsed = time.perf_counter() - t0
-    logger.info("Fold %d: wrote %d predictions to %s in %.1fs", fold_idx, n_preds, args.output, elapsed)
-
-
-def main_fold() -> None:
-    run_fold()
-
-
-# ---------------------------------------------------------------------------
-# Fold inference from pre-computed tensor cache  (dnn_fold_inference_from_cache)
-# ---------------------------------------------------------------------------
-
-
-def _load_tensor_shard(shard_path: str) -> dict:
-    """Load a single tensor shard and convert torch tensors back to numpy for GPU inference.
-
-    Supports both uncompressed ``.pt`` and gzip-compressed ``.pt.gz`` files.
-    Uses isal (ISA-L) for fast gzip decompression when available.
-    """
-    _torch = importlib.import_module("torch")
-
-    if shard_path.endswith(".gz"):
-        with igzip.open(shard_path, "rb") as f:
-            chunk = _torch.load(io.BytesIO(f.read()), map_location="cpu", weights_only=False)
-    else:
-        chunk = _torch.load(shard_path, map_location="cpu", weights_only=False)
-    result = {}
-    for key, val in chunk.items():
-        if isinstance(val, _torch.Tensor):
-            result[key] = val.numpy()
-        else:
-            result[key] = val
-    return result
-
-
-def _cpu_prep_worker_fn(
-    raw_queue: Queue,
-    gpu_queue: Queue,
-    error_event: Event,
-    error_holder: list[BaseException],
-) -> None:
-    """Stage 2: prepare GPU-ready arrays from raw tensor chunks."""
-    try:
-        while True:
-            chunk = raw_queue.get()
-            if chunk is None:
-                break
-            x_num = _compose_x_num(chunk)
-            prepared = {
-                "read_base_idx": _to_numpy_long(chunk["read_base_idx"]),
-                "ref_base_idx": _to_numpy_long(chunk["ref_base_idx"]),
-                "x_num": x_num,
-                "mask": _to_numpy(chunk["mask"]),
-            }
-            for key in ("tm_idx", "st_idx", "et_idx"):
-                if key in chunk:
-                    prepared[key] = _to_numpy_long(chunk[key])
-            prepared["_chrom"] = chunk["chrom"]
-            prepared["_pos"] = chunk["pos"]
-            prepared["_rn"] = chunk["rn"]
-            gpu_queue.put(prepared)
-    except Exception as exc:
-        logger.error("_cpu_prep_worker failed: %s", exc, exc_info=True)
-        error_holder.append(exc)
-        error_event.set()
-    finally:
-        gpu_queue.put(None)
-
-
-def _gpu_worker_fn(
-    engine,
-    gpu_queue: Queue,
-    result_queue: Queue,
-    batch_size: int,
-    error_event: Event,
-    error_holder: list[BaseException],
-) -> None:
-    """Stage 3: GPU inference only -- no CPU prep work."""
-    engine.push_context()
-    try:
-        while True:
-            prepared = gpu_queue.get()
-            if prepared is None:
-                break
-            n = prepared["x_num"].shape[0]
-            all_probs = []
-            for batch_start in range(0, n, batch_size):
-                s = slice(batch_start, min(batch_start + batch_size, n))
-                batch = {
-                    "read_base_idx": prepared["read_base_idx"][s],
-                    "ref_base_idx": prepared["ref_base_idx"][s],
-                    "x_num": prepared["x_num"][s],
-                    "mask": prepared["mask"][s],
-                }
-                for key in ("tm_idx", "st_idx", "et_idx"):
-                    if key in prepared:
-                        batch[key] = prepared[key][s]
-                all_probs.append(engine.predict_batch_prepared(batch))
-            result_queue.put((prepared["_chrom"], prepared["_pos"], prepared["_rn"], np.concatenate(all_probs)))
-            gpu_queue.task_done()
-    except Exception as exc:
-        logger.error("_gpu_worker failed: %s", exc, exc_info=True)
-        error_holder.append(exc)
-        error_event.set()
-    finally:
-        engine.pop_context()
-        result_queue.put(None)
-
-
-def _shard_loader_thread_fn(
-    shard_files: list,
-    raw_queue: Queue,
-    num_loader_threads: int,
-    error_event: Event,
-    error_holder: list[BaseException],
-) -> None:
-    """Stage 1: load tensor shards from disk into *raw_queue*."""
-    try:
-        with ThreadPoolExecutor(max_workers=num_loader_threads) as loader_pool:
-            max_pending = raw_queue.maxsize
-            shard_iter = iter(shard_files)
-            pending_loads = {
-                loader_pool.submit(_load_tensor_shard, str(p)) for p in itertools.islice(shard_iter, max_pending)
-            }
-            while pending_loads:
-                done_loads, pending_loads = wait(pending_loads, return_when=FIRST_COMPLETED)
-                for fut in done_loads:
-                    raw_queue.put(fut.result())
-                for p in itertools.islice(shard_iter, len(done_loads)):
-                    pending_loads.add(loader_pool.submit(_load_tensor_shard, str(p)))
-    except Exception as exc:
-        logger.error("_loader_thread failed: %s", exc, exc_info=True)
-        error_holder.append(exc)
-        error_event.set()
-    finally:
-        raw_queue.put(None)
-
-
-def _collect_cache_results(
-    result_queue: Queue,
-    total_shards: int,
-    error_event: Event,
-    t_start: float,
-    *,
-    streaming: bool,
-    writer=None,
-    predictions: dict[tuple[str, int, str], float] | None = None,
-) -> tuple[int, int]:
-    """Collect GPU results from *result_queue*, returning (total_preds, collected)."""
-    total_preds = 0
-    collected = 0
-    while collected < total_shards:
-        try:
-            item = result_queue.get(timeout=5.0)
-        except Empty:
-            if error_event.is_set():
-                break
-            continue
-        if item is None:
-            break
-        chroms, positions, rns, probs = item
-        n = len(probs)
-        if streaming:
-            table = pa.table(
-                {
-                    "chrom": chroms[:n].tolist() if hasattr(chroms, "tolist") else list(chroms[:n]),
-                    "pos": positions[:n].tolist() if hasattr(positions, "tolist") else list(positions[:n]),
-                    "rn": rns[:n].tolist() if hasattr(rns, "tolist") else list(rns[:n]),
-                    "prob": probs.tolist(),
-                }
-            )
-            writer.write_table(table)
-        else:
-            for j in range(n):
-                predictions[(str(chroms[j]), int(positions[j]), str(rns[j]))] = float(probs[j])
-        total_preds += n
-        collected += 1
-        if collected == 1 or collected % 10 == 0 or collected == total_shards:
-            logger.info(
-                "Cache inference progress: %d/%d shards (%.1fs)",
-                collected,
-                total_shards,
-                time.perf_counter() - t_start,
-            )
-    return total_preds, collected
-
-
-def _run_fold_inference_from_cache(
-    cache_dir: str,
-    engines: list,
-    batch_size: int = 512,
-    output_path: str | None = None,
-    num_loader_threads: int = 4,
-) -> dict[tuple[str, int, str], float] | int:
-    """Run GPU inference from pre-computed tensor cache shards.
-
-    Reads shard_*.pt files from *cache_dir* and feeds them directly to
-    GPU engines.  No CRAM access or tensorization needed -- all heavy
-    CPU work was done during the tensor cache creation step.
-
-    Uses a thread pool to prefetch shards from disk while GPUs process
-    current ones, overlapping I/O with computation.
-
-    Returns predictions dict or count (if *output_path* streaming).
-    """
-    cache_path = Path(cache_dir)
-    shard_files = sorted(cache_path.glob("shard_*.pt.gz")) or sorted(cache_path.glob("shard_*.pt"))
-    if not shard_files:
-        raise FileNotFoundError(f"No shard_*.pt files found in {cache_dir}")
-
-    total_shards = len(shard_files)
-    logger.info("Loading %d tensor shards from %s (%d loader threads)", total_shards, cache_dir, num_loader_threads)
-
-    # 3-stage pipeline queues and error tracking
-    raw_queue: Queue = Queue(maxsize=num_loader_threads + 1)
-    gpu_queue: Queue = Queue(maxsize=len(engines) * 2)
-    result_queue: Queue = Queue()
-    error_event = Event()
-    error_holder: list[BaseException] = []
-
-    engine = engines[0]
-
-    # Start pipeline threads
-    prep_thread = Thread(
-        target=_cpu_prep_worker_fn, args=(raw_queue, gpu_queue, error_event, error_holder), daemon=True
-    )
-    prep_thread.start()
-
-    gpu_thread = Thread(
-        target=_gpu_worker_fn,
-        args=(engine, gpu_queue, result_queue, batch_size, error_event, error_holder),
-        daemon=True,
-    )
-    gpu_thread.start()
-
-    loader_t = Thread(
-        target=_shard_loader_thread_fn,
-        args=(shard_files, raw_queue, num_loader_threads, error_event, error_holder),
-        daemon=True,
-    )
-    loader_t.start()
-
-    # Streaming output setup
-    streaming = output_path is not None
-    writer = None
-    if streaming:
-        schema = pa.schema([("chrom", pa.string()), ("pos", pa.int64()), ("rn", pa.string()), ("prob", pa.float64())])
-        writer = pq.ParquetWriter(output_path, schema)
-
-    predictions: dict[tuple[str, int, str], float] = {}
-    t_start = time.perf_counter()
-
-    try:
-        total_preds, _collected = _collect_cache_results(
-            result_queue,
-            total_shards,
-            error_event,
-            t_start,
-            streaming=streaming,
-            writer=writer,
-            predictions=predictions,
-        )
-
-        if error_event.is_set() and error_holder:
-            raise RuntimeError(f"Worker thread failed: {error_holder[0]}") from error_holder[0]
-
-        loader_t.join(timeout=30.0)
-        prep_thread.join(timeout=30.0)
-        gpu_thread.join(timeout=30.0)
-    finally:
-        if writer is not None:
-            writer.close()
-
-    elapsed = time.perf_counter() - t_start
-    logger.info("Cache inference complete: %d predictions in %.1fs", total_preds, elapsed)
-    return total_preds if streaming else predictions
-
-
-def _parse_fold_cache_args(argv: list[str] | None = None) -> argparse.Namespace:
-    ap = argparse.ArgumentParser(
-        description="DNN fold inference from pre-computed tensor cache (no CRAM access needed)",
-    )
-    ap.add_argument("--tensor-cache", required=True, help="Directory with shard_*.pt files")
-    ap.add_argument("--fold-metadata", required=True, help="Path to this fold's srsnv_dnn_metadata.json")
-    ap.add_argument("--output", required=True, help="Output predictions parquet path")
-    ap.add_argument("--backend", default="trt", choices=["trt", "pytorch"], help="Inference backend")
-    ap.add_argument("--gpus", default=None, help="Comma-separated GPU IDs (default: all visible)")
-    ap.add_argument("--batch-size", type=int, default=512)
-    return ap.parse_args(argv)
-
-
-def run_fold_from_cache(argv: list[str] | None = None) -> None:
-    """Run inference for a single fold from a pre-computed tensor cache."""
-    torch_mp = importlib.import_module("torch.multiprocessing")
-
-    torch_mp.set_sharing_strategy("file_system")
-
-    if argv is None:
-        argv = sys.argv[1:]
-    args = _parse_fold_cache_args(argv)
-
-    t0 = time.perf_counter()
-
-    # GPU setup
-    gpu_ids = None
-    if args.gpus:
-        gpu_ids = [int(g.strip()) for g in args.gpus.split(",")]
-
-    if gpu_ids is None:
-        _torch = importlib.import_module("torch")
-
-        gpu_ids = list(range(_torch.cuda.device_count())) if _torch.cuda.is_available() else [0]
-
-    # Load model engines
-    engines = _create_engines(args.fold_metadata, backend=args.backend, gpu_ids=gpu_ids)
-    logger.info("Loaded %d %s engine(s)", len(engines), args.backend)
-
-    try:
-        n_preds = _run_fold_inference_from_cache(
-            cache_dir=args.tensor_cache,
-            engines=engines,
-            batch_size=args.batch_size,
-            output_path=args.output,
-        )
-    finally:
-        for eng in engines:
-            eng.close()
-
-    elapsed = time.perf_counter() - t0
-    logger.info("Fold from cache: wrote %d predictions to %s in %.1fs", n_preds, args.output, elapsed)
-
-
-def main_fold_from_cache() -> None:
-    run_fold_from_cache()
-
-
-# ---------------------------------------------------------------------------
-# Merge + annotate CLI  (dnn_merge_and_annotate)
-# ---------------------------------------------------------------------------
-
-
-def _parse_merge_args(argv: list[str] | None = None) -> argparse.Namespace:
-    ap = argparse.ArgumentParser(
-        description="Merge per-fold DNN predictions and annotate VCF",
-    )
-    ap.add_argument("--featuremap-vcf", required=True, help="Input featuremap VCF to annotate")
-    ap.add_argument("--fold-predictions", nargs="+", required=True, help="Per-fold prediction parquet files")
-    ap.add_argument("--fold-metadata", required=True, help="Fold 0 metadata JSON (for quality recalibration LUT)")
-    ap.add_argument("--output", required=True, help="Output annotated VCF path")
-    ap.add_argument("--low-qual-threshold", type=float, default=40.0, help="SNVQ threshold for PASS filter")
-    ap.add_argument("--process-number", type=int, default=-2, help="Parallel processes for VCF annotation")
-    return ap.parse_args(argv)
-
-
-def _load_contig_predictions(
-    contig: str,
-    pred_paths: list[str],
-) -> tuple[np.ndarray, list[str], list[float]] | None:
-    """Load and sort predictions for one contig from all fold parquet files.
-
-    Returns ``(pos_arr, rn_list, prob_list)`` sorted by position, or *None*
-    if no predictions exist for *contig*.
-    """
-    all_pos: list[np.ndarray] = []
-    all_rn: list[str] = []
-    all_prob: list[np.ndarray] = []
-    for path in pred_paths:
-        table = pq.read_table(path, filters=[("chrom", "==", contig)])
-        if table.num_rows == 0:
-            continue
-        all_pos.append(table.column("pos").to_numpy())
-        all_rn.extend(table.column("rn").to_pylist())
-        all_prob.append(table.column("prob").to_numpy())
-
-    if not all_pos:
-        return None
-
-    pos_arr = np.concatenate(all_pos)
-    prob_arr = np.concatenate(all_prob)
-    del all_pos, all_prob
-
-    order = np.argsort(pos_arr, kind="stable")
-    pos_arr = pos_arr[order]
-    prob_list = prob_arr[order].tolist()
-    rn_list = [all_rn[i] for i in order]
-    return pos_arr, rn_list, prob_list
-
-
-def _build_snvq_lut(
-    quality_lut_x: list[float] | None,
-    quality_lut_y: list[float] | None,
-    max_phred: float,
-) -> dict[int, float]:
-    """Pre-compute MQUAL->SNVQ lookup dict keyed by int(mqual*100)."""
-    if quality_lut_x is None or quality_lut_y is None:
-        return {}
-    lut_x = np.asarray(quality_lut_x, dtype=np.float64)
-    lut_y = np.asarray(quality_lut_y, dtype=np.float64)
-    right = float(lut_y[-1])
-    snvq_lut: dict[int, float] = {}
-    for mq_cent in range(int(max_phred * 100) + 1):
-        sq = float(np.interp(mq_cent / 100.0, lut_x, lut_y, left=0.0, right=right))
-        snvq_lut[mq_cent] = round(sq, 2)
-    return snvq_lut
-
-
-def _annotate_record(
-    rec: pysam.VariantRecord,
-    preds_at_pos: dict[str, float],
-    snvq_lut: dict[int, float],
-    max_phred: float,
-    min_prob_error: float,
-    low_qual_threshold: float,
-) -> None:
-    """Annotate a single VCF record with MQUAL/SNVQ from matched predictions."""
-    try:
-        rn_val = rec.samples[0].get("RN")
-    except (KeyError, IndexError):
-        rn_val = None
-
-    if not rn_val or not preds_at_pos:
-        return
-
-    rns = rn_val if isinstance(rn_val, tuple) else (rn_val,)
-
-    # Compute per-read MQUAL
-    mquals = []
-    for rn in rns:
-        prob = preds_at_pos.get(rn if isinstance(rn, str) else str(rn))
-        prob = prob if prob is not None and prob > 0 else 0.0
-        if prob > 0:
-            mq = -10.0 * math.log10(max(1.0 - prob, min_prob_error))
-            mq = min(mq, max_phred)
-        else:
-            mq = 0.0
-        mquals.append(round(mq, 2))
-
-    # Compute SNVQ
-    if snvq_lut:
-        snvqs = [snvq_lut.get(int(mq * 100), mq) for mq in mquals]
-    else:
-        snvqs = list(mquals)
-
-    rec.samples[0]["MQUAL"] = tuple(mquals)
-    rec.samples[0]["SNVQ"] = tuple(snvqs)
-
-    max_snvq = max(snvqs)
-    rec.qual = max_snvq
-    if max_snvq >= low_qual_threshold:
-        rec.filter.add("PASS")
-    else:
-        rec.filter.add("LowQual")
-
-
-def _copy_contig_records(vcf_in: str, vcf_out: str, contig: str) -> str:
-    """Copy contig records through without annotation (no predictions available)."""
-    with pysam.VariantFile(vcf_in) as inp:
-        try:
-            inp.header.add_line('##FILTER=<ID=LowQual,Description="SNVQ below quality threshold">')
-        except ValueError:
-            pass
-        with pysam.VariantFile(vcf_out, "w", header=inp.header) as out:
-            for rec in inp.fetch(contig):
-                out.write(rec)
-    pysam.tabix_index(vcf_out, preset="vcf", force=True)
-    return vcf_out
-
-
-def _merge_contig_worker(
-    contig: str,
-    vcf_in: str,
-    vcf_out: str,
-    pred_paths: list[str],
-    quality_lut_x: list[float] | None,
-    quality_lut_y: list[float] | None,
-    low_qual_threshold: float,
-) -> str:
-    """Annotate one contig using streaming merge-join.
-
-    Each worker independently reads its predictions from parquet (predicate
-    pushdown) and streams VCF records.  Both sources are sorted by position,
-    enabling an O(N+M) merge cursor instead of hash-table lookups.
-
-    No shared state with the parent -- fully independent, picklable args only.
-    """
-    loaded = _load_contig_predictions(contig, pred_paths)
-    if loaded is None:
-        return _copy_contig_records(vcf_in, vcf_out, contig)
-
-    pos_arr, rn_list, prob_list = loaded
-    n_preds = len(pos_arr)
-    max_phred = float(MAX_PHRED)
-    min_prob_error = 10.0 ** (-max_phred / 10.0)
-    snvq_lut = _build_snvq_lut(quality_lut_x, quality_lut_y, max_phred)
-
-    # Stream VCF records with merge cursor
-    cursor = 0
-    with pysam.VariantFile(vcf_in) as inp:
-        try:
-            inp.header.add_line('##FILTER=<ID=LowQual,Description="SNVQ below quality threshold">')
-        except ValueError:
-            pass
-
-        with pysam.VariantFile(vcf_out, "w", header=inp.header) as out:
-            for rec in inp.fetch(contig):
-                pos = rec.pos
-
-                while cursor < n_preds and pos_arr[cursor] < pos:
-                    cursor += 1
-
-                preds_at_pos: dict[str, float] = {}
-                i = cursor
-                while i < n_preds and pos_arr[i] == pos:
-                    preds_at_pos[rn_list[i]] = prob_list[i]
-                    i += 1
-
-                _annotate_record(rec, preds_at_pos, snvq_lut, max_phred, min_prob_error, low_qual_threshold)
-                out.write(rec)
-
-    pysam.tabix_index(vcf_out, preset="vcf", force=True)
-    return vcf_out
-
-
-def run_merge(argv: list[str] | None = None) -> None:
-    """Merge per-fold predictions and annotate the featuremap VCF.
-
-    Uses a streaming merge-join approach: each contig is processed by an
-    independent subprocess that reads its own predictions from parquet
-    (with predicate pushdown) and streams VCF records.  Both are sorted
-    by position, enabling an O(N+M) merge cursor.
-
-    No data sharing between processes — each worker is fully self-contained.
-    """
-    if argv is None:
-        argv = sys.argv[1:]
-    args = _parse_merge_args(argv)
-
-    t0 = time.perf_counter()
-
-    # Extract quality LUT as plain lists (picklable, tiny)
-    with open(args.fold_metadata) as f:
-        metadata = json.load(f)
-    lut = metadata.get("quality_recalibration_table")
-    min_lut_entries = 2
-    quality_lut_x = lut[0] if lut and len(lut) >= min_lut_entries else None
-    quality_lut_y = lut[1] if lut and len(lut) >= min_lut_entries else None
-
-    # Quick scan: count predictions per contig (chrom column only — fast)
-    contig_counts: dict[str, int] = {}
-    total_preds = 0
-    for pred_path in args.fold_predictions:
-        chrom_col = pq.read_table(pred_path, columns=["chrom"]).column("chrom")
-        for entry in pc.value_counts(chrom_col).to_pylist():
-            c = str(entry["values"])
-            contig_counts[c] = contig_counts.get(c, 0) + entry["counts"]
-            total_preds += entry["counts"]
-
-    logger.info(
-        "Found %d predictions across %d contigs from %d fold files",
-        total_preds,
-        len(contig_counts),
-        len(args.fold_predictions),
-    )
-
-    # Find non-empty contigs in VCF
-    out_dir = os.path.dirname(args.output) or "."
-
-    with pysam.VariantFile(args.featuremap_vcf) as input_vcf:
-        contig_tasks = []
-        skipped = 0
-        for contig in input_vcf.header.contigs:
-            if contig not in contig_counts:
-                skipped += 1
-                continue
-            try:
-                next(input_vcf.fetch(contig))
-            except StopIteration:
-                skipped += 1
-                continue
-            contig_tasks.append(contig)
-
-    if skipped:
-        logger.info("Skipped %d empty contigs", skipped)
-
-    # Schedule largest contigs first for better load balancing
-    contig_tasks.sort(key=lambda c: contig_counts.get(c, 0), reverse=True)
-
-    max_cpus = args.process_number if args.process_number > 0 else (os.cpu_count() or 4)
-    num_workers = min(max_cpus, max(1, len(contig_tasks)))
-    logger.info(
-        "Annotating %d contigs with %d processes (largest: %s with %d predictions)",
-        len(contig_tasks),
-        num_workers,
-        contig_tasks[0] if contig_tasks else "?",
-        contig_counts.get(contig_tasks[0], 0) if contig_tasks else 0,
-    )
-
-    # Each worker independently reads parquet + VCF — no shared state, no pickle of
-    # large data, no COW issues.  Only small picklable args are sent to each worker.
-    tmp_output_paths: list[str] = []
-    with ProcessPoolExecutor(max_workers=num_workers) as pool:
-        futures = {}
-        for contig in contig_tasks:
-            out_path = os.path.join(out_dir, contig + ".vcf.gz")
-            fut = pool.submit(
-                _merge_contig_worker,
-                contig=contig,
-                vcf_in=args.featuremap_vcf,
-                vcf_out=out_path,
-                pred_paths=list(args.fold_predictions),
-                quality_lut_x=quality_lut_x,
-                quality_lut_y=quality_lut_y,
-                low_qual_threshold=args.low_qual_threshold,
-            )
-            futures[fut] = (contig, out_path)
-
-        for fut, (contig, out_path) in futures.items():
-            fut.result()  # raise on error
-            tmp_output_paths.append(out_path)
-            logger.info("Annotated contig %s", contig)
-
-    VcfAnnotator.merge_temp_files(tmp_output_paths, args.output, process_number=1)
-
-    elapsed = time.perf_counter() - t0
-    logger.info("Merge + annotate complete: %d predictions -> %s in %.1fs", total_preds, args.output, elapsed)
-
-
-def main_merge() -> None:
-    run_merge()
-
-
-if __name__ == "__main__":
-    main()
